fgweb_project_api/requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
django-db-connection-pool==1.2.5
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
mysqlclient==2.2.4
pillow==10.4.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-decouple==3.8
redis==5.0.8
SQLAlchemy==2.0.32
sqlparams==6.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
